refactor: fixing project tree: application layer deleted
